// src/generated/FrsPageReqIdl.ts
export interface FrsPageReqIdl {
  data: FrsPageReqIdl_DataReq | undefined;
}

export interface FrsPageReqIdl_DataReq {
  common: CommonReq | undefined;
  kw: string;
  rn: number;
  rnNeed: number;
  isGood: number;
  cid: number;
  pn: number;
  sortType: number;
}

function createBaseFrsPageReqIdl(): FrsPageReqIdl {
  return { data: undefined };
}

export const FrsPageReqIdl: MessageFns<FrsPageReqIdl> = {
  encode(message: FrsPageReqIdl, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.data !== undefined) {
      FrsPageReqIdl_DataReq.encode(message.data, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): FrsPageReqIdl {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseFrsPageReqIdl();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.data = FrsPageReqIdl_DataReq.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): FrsPageReqIdl {
    return { data: isSet(object.data) ? FrsPageReqIdl_DataReq.fromJSON(object.data) : undefined };
  },

  toJSON(message: FrsPageReqIdl): unknown {
    const obj: any = {};
    if (message.data !== undefined) {
      obj.data = FrsPageReqIdl_DataReq.toJSON(message.data);
    }
    return obj;
  },

  create(base?: DeepPartial<FrsPageReqIdl>): FrsPageReqIdl {
    return FrsPageReqIdl.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<FrsPageReqIdl>): FrsPageReqIdl {
    const message = createBaseFrsPageReqIdl();
    message.data = (object.data !== undefined && object.data !== null)
      ? FrsPageReqIdl_DataReq.fromPartial(object.data)
      : undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}