// src/generated/FrsPageResIdl.ts
export interface FrsPageResIdl {
  error: Error | undefined;
  data: FrsPageResIdl_DataRes | undefined;
}

export interface FrsPageResIdl_DataRes {
  forum: FrsPageResIdl_DataRes_ForumInfo | undefined;
  page: Page | undefined;
  threadList: ThreadInfo[];
  userList: User[];
  navTabInfo: FrsPageResIdl_DataRes_NavTabInfo | undefined;
  forumRule: FrsPageResIdl_DataRes_ForumRuleStatus | undefined;
}

export interface FrsPageResIdl_DataRes_ForumInfo {
  id: string;
  name: string;
  firstClass: string;
  secondClass: string;
  memberNum: number;
  threadNum: number;
  postNum: number;
  managers: FrsPageResIdl_DataRes_ForumInfo_Manager[];
}

export interface FrsPageResIdl_DataRes_ForumInfo_Manager {
}

export interface FrsPageResIdl_DataRes_NavTabInfo {
  tab: FrsTabInfo[];
}

export interface FrsPageResIdl_DataRes_ForumRuleStatus {
  hasForumRule: number;
}

function createBaseFrsPageResIdl(): FrsPageResIdl {
  return { error: undefined, data: undefined };
}

export const FrsPageResIdl: MessageFns<FrsPageResIdl> = {
  encode(message: FrsPageResIdl, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.error !== undefined) {
      Error.encode(message.error, writer.uint32(10).fork()).join();
    }
    if (message.data !== undefined) {
      FrsPageResIdl_DataRes.encode(message.data, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): FrsPageResIdl {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseFrsPageResIdl();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.error = Error.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.data = FrsPageResIdl_DataRes.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): FrsPageResIdl {
    return {
      error: isSet(object.error) ? Error.fromJSON(object.error) : undefined,
      data: isSet(object.data) ? FrsPageResIdl_DataRes.fromJSON(object.data) : undefined,
    };
  },

  toJSON(message: FrsPageResIdl): unknown {
    const obj: any = {};
    if (message.error !== undefined) {
      obj.error = Error.toJSON(message.error);
    }
    if (message.data !== undefined) {
      obj.data = FrsPageResIdl_DataRes.toJSON(message.data);
    }
    return obj;
  },

  create(base?: DeepPartial<FrsPageResIdl>): FrsPageResIdl {
    return FrsPageResIdl.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<FrsPageResIdl>): FrsPageResIdl {
    const message = createBaseFrsPageResIdl();
    message.error = (object.error !== undefined && object.error !== null) ? Error.fromPartial(object.error) : undefined;
    message.data = (object.data !== undefined && object.data !== null)
      ? FrsPageResIdl_DataRes.fromPartial(object.data)
      : undefined;
    return message;
  },
};
// ...
export const FrsPageResIdl_DataRes_NavTabInfo: MessageFns<FrsPageResIdl_DataRes_NavTabInfo> = {
  encode(message: FrsPageResIdl_DataRes_NavTabInfo, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    for (const v of message.tab) {
      FrsTabInfo.encode(v!, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): FrsPageResIdl_DataRes_NavTabInfo {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseFrsPageResIdl_DataRes_NavTabInfo();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.tab.push(FrsTabInfo.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): FrsPageResIdl_DataRes_NavTabInfo {
    return { tab: globalThis.Array.isArray(object?.tab) ? object.tab.map((e: any) => FrsTabInfo.fromJSON(e)) : [] };
  },

  toJSON(message: FrsPageResIdl_DataRes_NavTabInfo): unknown {
    const obj: any = {};
    if (message.tab?.length) {
      obj.tab = message.tab.map((e) => FrsTabInfo.toJSON(e));
    }
    return obj;
  },

  create(base?: DeepPartial<FrsPageResIdl_DataRes_NavTabInfo>): FrsPageResIdl_DataRes_NavTabInfo {
    return FrsPageResIdl_DataRes_NavTabInfo.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<FrsPageResIdl_DataRes_NavTabInfo>): FrsPageResIdl_DataRes_NavTabInfo {
    const message = createBaseFrsPageResIdl_DataRes_NavTabInfo();
    message.tab = object.tab?.map((e) => FrsTabInfo.fromPartial(e)) || [];
    return message;
  },
};

function createBaseFrsPageResIdl_DataRes_ForumRuleStatus(): FrsPageResIdl_DataRes_ForumRuleStatus {
  return { hasForumRule: 0 };
}

export const FrsPageResIdl_DataRes_ForumRuleStatus: MessageFns<FrsPageResIdl_DataRes_ForumRuleStatus> = {
  encode(message: FrsPageResIdl_DataRes_ForumRuleStatus, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.hasForumRule !== 0) {
      writer.uint32(32).int32(message.hasForumRule);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): FrsPageResIdl_DataRes_ForumRuleStatus {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseFrsPageResIdl_DataRes_ForumRuleStatus();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.hasForumRule = reader.int32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): FrsPageResIdl_DataRes_ForumRuleStatus {
    return {
      hasForumRule: isSet(object.hasForumRule)
        ? globalThis.Number(object.hasForumRule)
        : isSet(object.has_forum_rule)
        ? globalThis.Number(object.has_forum_rule)
        : 0,
    };
  },

  toJSON(message: FrsPageResIdl_DataRes_ForumRuleStatus): unknown {
    const obj: any = {};
    if (message.hasForumRule !== 0) {
      obj.hasForumRule = Math.round(message.hasForumRule);
    }
    return obj;
  },

  create(base?: DeepPartial<FrsPageResIdl_DataRes_ForumRuleStatus>): FrsPageResIdl_DataRes_ForumRuleStatus {
    return FrsPageResIdl_DataRes_ForumRuleStatus.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<FrsPageResIdl_DataRes_ForumRuleStatus>): FrsPageResIdl_DataRes_ForumRuleStatus {
    const message = createBaseFrsPageResIdl_DataRes_ForumRuleStatus();
    message.hasForumRule = object.hasForumRule ?? 0;
    return message;
  },
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}