// src/generated/GetForumDetailReqIdl.ts
export interface GetForumDetailReqIdl {
  data: GetForumDetailReqIdl_DataReq | undefined;
}

export interface GetForumDetailReqIdl_DataReq {
  forumId: string;
  common: CommonReq | undefined;
}

function createBaseGetForumDetailReqIdl(): GetForumDetailReqIdl {
  return { data: undefined };
}

export const GetForumDetailReqIdl: MessageFns<GetForumDetailReqIdl> = {
  encode(message: GetForumDetailReqIdl, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.data !== undefined) {
      GetForumDetailReqIdl_DataReq.encode(message.data, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): GetForumDetailReqIdl {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetForumDetailReqIdl();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.data = GetForumDetailReqIdl_DataReq.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GetForumDetailReqIdl {
    return { data: isSet(object.data) ? GetForumDetailReqIdl_DataReq.fromJSON(object.data) : undefined };
  },

  toJSON(message: GetForumDetailReqIdl): unknown {
    const obj: any = {};
    if (message.data !== undefined) {
      obj.data = GetForumDetailReqIdl_DataReq.toJSON(message.data);
    }
    return obj;
  },

  create(base?: DeepPartial<GetForumDetailReqIdl>): GetForumDetailReqIdl {
    return GetForumDetailReqIdl.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<GetForumDetailReqIdl>): GetForumDetailReqIdl {
    const message = createBaseGetForumDetailReqIdl();
    message.data = (object.data !== undefined && object.data !== null)
      ? GetForumDetailReqIdl_DataReq.fromPartial(object.data)
      : undefined;
    return message;
  },
};

function createBaseGetForumDetailReqIdl_DataReq(): GetForumDetailReqIdl_DataReq {
  return { forumId: "0", common: undefined };
}

export const GetForumDetailReqIdl_DataReq: MessageFns<GetForumDetailReqIdl_DataReq> = {
  encode(message: GetForumDetailReqIdl_DataReq, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.forumId !== "0") {
      writer.uint32(8).int64(message.forumId);
    }
    if (message.common !== undefined) {
      CommonReq.encode(message.common, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): GetForumDetailReqIdl_DataReq {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetForumDetailReqIdl_DataReq();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.forumId = reader.int64().toString();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.common = CommonReq.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GetForumDetailReqIdl_DataReq {
    return {
      forumId: isSet(object.forumId)
        ? globalThis.String(object.forumId)
        : isSet(object.forum_id)
        ? globalThis.String(object.forum_id)
        : "0",
      common: isSet(object.common) ? CommonReq.fromJSON(object.common) : undefined,
    };
  },

  toJSON(message: GetForumDetailReqIdl_DataReq): unknown {
    const obj: any = {};
    if (message.forumId !== "0") {
      obj.forumId = message.forumId;
    }
    if (message.common !== undefined) {
      obj.common = CommonReq.toJSON(message.common);
    }
    return obj;
  },

  create(base?: DeepPartial<GetForumDetailReqIdl_DataReq>): GetForumDetailReqIdl_DataReq {
    return GetForumDetailReqIdl_DataReq.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<GetForumDetailReqIdl_DataReq>): GetForumDetailReqIdl_DataReq {
    const message = createBaseGetForumDetailReqIdl_DataReq();
    message.forumId = object.forumId ?? "0";
    message.common = (object.common !== undefined && object.common !== null)
      ? CommonReq.fromPartial(object.common)
      : undefined;
    return message;
  },
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}