// src/generated/GetForumDetailResIdl.ts
export interface GetForumDetailResIdl {
  error: Error | undefined;
  data: GetForumDetailResIdl_DataRes | undefined;
}

export interface GetForumDetailResIdl_DataRes {
  forumInfo: GetForumDetailResIdl_DataRes_RecommendForumInfo | undefined;
  electionTab: GetForumDetailResIdl_DataRes_ManagerElectionTab | undefined;
}

export interface GetForumDetailResIdl_DataRes_RecommendForumInfo {
  avatar: string;
  forumId: string;
  forumName: string;
  memberCount: number;
  threadCount: number;
  slogan: string;
  lv1Name: string;
  avatarOrigin: string;
}

export interface GetForumDetailResIdl_DataRes_ManagerElectionTab {
  newStrategyText: string;
}

function createBaseGetForumDetailResIdl(): GetForumDetailResIdl {
  return { error: undefined, data: undefined };
}

export const GetForumDetailResIdl: MessageFns<GetForumDetailResIdl> = {
  encode(message: GetForumDetailResIdl, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.error !== undefined) {
      Error.encode(message.error, writer.uint32(10).fork()).join();
    }
    if (message.data !== undefined) {
      GetForumDetailResIdl_DataRes.encode(message.data, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): GetForumDetailResIdl {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetForumDetailResIdl();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.error = Error.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.data = GetForumDetailResIdl_DataRes.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GetForumDetailResIdl {
    return {
      error: isSet(object.error) ? Error.fromJSON(object.error) : undefined,
      data: isSet(object.data) ? GetForumDetailResIdl_DataRes.fromJSON(object.data) : undefined,
    };
  },

  toJSON(message: GetForumDetailResIdl): unknown {
    const obj: any = {};
    if (message.error !== undefined) {
      obj.error = Error.toJSON(message.error);
    }
    if (message.data !== undefined) {
      obj.data = GetForumDetailResIdl_DataRes.toJSON(message.data);
    }
    return obj;
  },

  create(base?: DeepPartial<GetForumDetailResIdl>): GetForumDetailResIdl {
    return GetForumDetailResIdl.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<GetForumDetailResIdl>): GetForumDetailResIdl {
    const message = createBaseGetForumDetailResIdl();
    message.error = (object.error !== undefined && object.error !== null) ? Error.fromPartial(object.error) : undefined;
    message.data = (object.data !== undefined && object.data !== null)
      ? GetForumDetailResIdl_DataRes.fromPartial(object.data)
      : undefined;
    return message;
  },
};

function createBaseGetForumDetailResIdl_DataRes(): GetForumDetailResIdl_DataRes {
  return { forumInfo: undefined, electionTab: undefined };
}

export const GetForumDetailResIdl_DataRes: MessageFns<GetForumDetailResIdl_DataRes> = {
  encode(message: GetForumDetailResIdl_DataRes, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.forumInfo !== undefined) {
      GetForumDetailResIdl_DataRes_RecommendForumInfo.encode(message.forumInfo, writer.uint32(10).fork()).join();
    }
    if (message.electionTab !== undefined) {
      GetForumDetailResIdl_DataRes_ManagerElectionTab.encode(message.electionTab, writer.uint32(66).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): GetForumDetailResIdl_DataRes {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetForumDetailResIdl_DataRes();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.forumInfo = GetForumDetailResIdl_DataRes_RecommendForumInfo.decode(reader, reader.uint32());
          continue;
        }
        case 8: {
          if (tag !== 66) {
            break;
          }

          message.electionTab = GetForumDetailResIdl_DataRes_ManagerElectionTab.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GetForumDetailResIdl_DataRes {
    return {
      forumInfo: isSet(object.forumInfo)
        ? GetForumDetailResIdl_DataRes_RecommendForumInfo.fromJSON(object.forumInfo)
        : isSet(object.forum_info)
        ? GetForumDetailResIdl_DataRes_RecommendForumInfo.fromJSON(object.forum_info)
        : undefined,
      electionTab: isSet(object.electionTab)
        ? GetForumDetailResIdl_DataRes_ManagerElectionTab.fromJSON(object.electionTab)
        : isSet(object.election_tab)
        ? GetForumDetailResIdl_DataRes_ManagerElectionTab.fromJSON(object.election_tab)
        : undefined,
    };
  },

  toJSON(message: GetForumDetailResIdl_DataRes): unknown {
    const obj: any = {};
    if (message.forumInfo !== undefined) {
      obj.forumInfo = GetForumDetailResIdl_DataRes_RecommendForumInfo.toJSON(message.forumInfo);
    }
    if (message.electionTab !== undefined) {
      obj.electionTab = GetForumDetailResIdl_DataRes_ManagerElectionTab.toJSON(message.electionTab);
    }
    return obj;
  },

  create(base?: DeepPartial<GetForumDetailResIdl_DataRes>): GetForumDetailResIdl_DataRes {
    return GetForumDetailResIdl_DataRes.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<GetForumDetailResIdl_DataRes>): GetForumDetailResIdl_DataRes {
    const message = createBaseGetForumDetailResIdl_DataRes();
    message.forumInfo = (object.forumInfo !== undefined && object.forumInfo !== null)
      ? GetForumDetailResIdl_DataRes_RecommendForumInfo.fromPartial(object.forumInfo)
      : undefined;
    message.electionTab = (object.electionTab !== undefined && object.electionTab !== null)
      ? GetForumDetailResIdl_DataRes_ManagerElectionTab.fromPartial(object.electionTab)
      : undefined;
    return message;
  },
};
// ...
export const GetForumDetailResIdl_DataRes_ManagerElectionTab: MessageFns<
  GetForumDetailResIdl_DataRes_ManagerElectionTab
> = {
  encode(
    message: GetForumDetailResIdl_DataRes_ManagerElectionTab,
    writer: BinaryWriter = new BinaryWriter(),
  ): BinaryWriter {
    if (message.newStrategyText !== "") {
      writer.uint32(42).string(message.newStrategyText);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): GetForumDetailResIdl_DataRes_ManagerElectionTab {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetForumDetailResIdl_DataRes_ManagerElectionTab();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.newStrategyText = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GetForumDetailResIdl_DataRes_ManagerElectionTab {
    return {
      newStrategyText: isSet(object.newStrategyText)
        ? globalThis.String(object.newStrategyText)
        : isSet(object.new_strategy_text)
        ? globalThis.String(object.new_strategy_text)
        : "",
    };
  },

  toJSON(message: GetForumDetailResIdl_DataRes_ManagerElectionTab): unknown {
    const obj: any = {};
    if (message.newStrategyText !== "") {
      obj.newStrategyText = message.newStrategyText;
    }
    return obj;
  },

  create(
    base?: DeepPartial<GetForumDetailResIdl_DataRes_ManagerElectionTab>,
  ): GetForumDetailResIdl_DataRes_ManagerElectionTab {
    return GetForumDetailResIdl_DataRes_ManagerElectionTab.fromPartial(base ?? {});
  },
  fromPartial(
    object: DeepPartial<GetForumDetailResIdl_DataRes_ManagerElectionTab>,
  ): GetForumDetailResIdl_DataRes_ManagerElectionTab {
    const message = createBaseGetForumDetailResIdl_DataRes_ManagerElectionTab();
    message.newStrategyText = object.newStrategyText ?? "";
    return message;
  },
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}